+++ src/classes/offset_datetime.rs
@@ -640,18 +640,6 @@ fn to_stdlib(cls: HeapType<OffsetDateTime>, slf: OffsetDateTime) -> PyReturn {
         .map(|owned| owned.map(|dt| dt.as_py_obj()))
 }
 
-fn py_datetime(cls: HeapType<OffsetDateTime>, slf: OffsetDateTime) -> PyReturn {
-    let &State {
-        warn_deprecation, ..
-    } = cls.state();
-    warn_with_class(
-        warn_deprecation,
-        c"py_datetime() is deprecated. Use to_stdlib() instead.",
-        1,
-    )?;
-    to_stdlib(cls, slf)
-}
-
 fn date(cls: HeapType<OffsetDateTime>, OffsetDateTime { date, .. }: OffsetDateTime) -> PyReturn {
     date.to_obj(cls.state().date_type)
 }
@@ -724,10 +712,7 @@ fn offset_stale_warning(state: &State, msg: &CStr) -> PyResult<()> {
     warn_with_class(state.warn_potentially_stale_offset, msg, 1)
 }
 
-/// Check for deprecated `ignore_dst` and new `stale_offset_ok`
-/// kwargs in a kwargs iterator that only has these optional kwargs remaining,
-/// and emit stale offset warning.
-fn check_ignore_dst_and_stale_offset(
+fn check_stale_offset(
     fname: &str,
     kwargs: &mut IterKwargs,
     state: &State,
@@ -735,9 +720,7 @@ fn check_ignore_dst_and_stale_offset(
 ) -> PyResult<()> {
     let mut suppress = false;
     handle_kwargs(fname, kwargs, |key, value, eq| {
-        if eq(key, state.str_ignore_dst) {
-            warn_with_class(state.warn_deprecation, doc::IGNORE_DST_DEPRECATED_MSG, 1)?;
-        } else if eq(key, state.str_stale_offset_ok) {
+        if eq(key, state.str_stale_offset_ok) {
             suppress = value.is_truthy();
         } else {
             return Ok(false);
@@ -757,12 +740,7 @@ fn replace_date(
     kwargs: &mut IterKwargs,
 ) -> PyReturn {
     let state = cls.state();
-    check_ignore_dst_and_stale_offset(
-        "replace_date",
-        kwargs,
-        state,
-        doc::OFFSET_REPLACE_STALE_MSG,
-    )?;
+    check_stale_offset("replace_date", kwargs, state, doc::OFFSET_REPLACE_STALE_MSG)?;
     let &[arg] = args else {
         raise_type_err("replace_date() takes exactly 1 positional argument")?
     };
@@ -782,12 +760,7 @@ fn replace_time(
     kwargs: &mut IterKwargs,
 ) -> PyReturn {
     let state = cls.state();
-    check_ignore_dst_and_stale_offset(
-        "replace_time",
-        kwargs,
-        state,
-        doc::OFFSET_REPLACE_STALE_MSG,
-    )?;
+    check_stale_offset("replace_time", kwargs, state, doc::OFFSET_REPLACE_STALE_MSG)?;
     let &[arg] = args else {
         raise_type_err("replace_time() takes exactly 1 positional argument")?
     };
@@ -839,7 +812,6 @@ fn replace(
     }
     let state = cls.state();
     let &State {
-        str_ignore_dst,
         str_offset,
         str_stale_offset_ok,
         str_year,
@@ -864,13 +836,10 @@ fn replace(
     let mut minute = time.minute.into();
     let mut second = time.second.into();
     let mut nanos = time.subsec.get() as _;
-    let mut got_ignore_dst = false;
     let mut suppress_stale = false;
 
     handle_kwargs("replace", kwargs, |key, value, eq| {
-        if eq(key, str_ignore_dst) {
-            got_ignore_dst = true;
-        } else if eq(key, str_stale_offset_ok) {
+        if eq(key, str_stale_offset_ok) {
             suppress_stale = value.is_truthy();
         } else if eq(key, str_offset) {
             offset = Offset::from_obj(value, time_delta_type)?;
@@ -898,9 +867,6 @@ fn replace(
         Ok(true)
     })?;
 
-    if got_ignore_dst {
-        warn_with_class(state.warn_deprecation, doc::IGNORE_DST_DEPRECATED_MSG, 1)?;
-    }
     if !suppress_stale {
         offset_stale_warning(state, doc::OFFSET_REPLACE_STALE_MSG)?;
     }
@@ -917,7 +883,7 @@ fn now(cls: HeapType<OffsetDateTime>, args: &[PyObj], kwargs: &mut IterKwargs) -
     let &[offset_obj] = args else {
         raise_type_err("now() takes exactly 1 positional argument")?
     };
-    check_ignore_dst_and_stale_offset("now", kwargs, state, doc::OFFSET_NOW_STALE_MSG)?;
+    check_stale_offset("now", kwargs, state, doc::OFFSET_NOW_STALE_MSG)?;
     let offset = Offset::from_obj(offset_obj, state.time_delta_type)?;
     state
         .time_ns()?
@@ -925,22 +891,6 @@ fn now(cls: HeapType<OffsetDateTime>, args: &[PyObj], kwargs: &mut IterKwargs) -
         // SAFETY: Exception types are safe to reference during Python runtime
         .ok_or_raise(unsafe { PyExc_OSError }, "Date is out of range")?
         .to_obj(cls)
-}
-
-fn from_py_datetime(cls: HeapType<OffsetDateTime>, arg: PyObj) -> PyReturn {
-    let &State {
-        warn_deprecation, ..
-    } = cls.state();
-    warn_with_class(
-        warn_deprecation,
-        c"from_py_datetime() is deprecated. Use OffsetDateTime() constructor instead.",
-        1,
-    )?;
-    if let Some(py_dt) = arg.cast_allow_subclass::<PyDateTime>() {
-        OffsetDateTime::from_py(py_dt)?.to_obj(cls)
-    } else {
-        raise_type_err("argument must be a datetime.datetime instance")?
-    }
 }
 
 pub(crate) fn to_plain(cls: HeapType<OffsetDateTime>, slf: OffsetDateTime) -> PyReturn {
@@ -988,7 +938,6 @@ fn shift_method(
     let fname = if negate { "subtract" } else { "add" };
     let state = cls.state();
     let &State {
-        str_ignore_dst,
         str_stale_offset_ok,
         str_years,
         str_months,
@@ -1001,25 +950,19 @@ fn shift_method(
         str_microseconds,
         str_nanoseconds,
         time_delta_type,
-        date_delta_type,
-        datetime_delta_type,
         itemized_date_delta_type,
         itemized_delta_type,
-        warn_deprecation,
         ..
     } = state;
     let mut months = DeltaMonths::ZERO;
     let mut days = DeltaDays::ZERO;
     let mut tdelta = TimeDelta::ZERO;
-    let mut got_ignore_dst = false;
     let mut suppress_stale = false;
 
     match *args {
         [arg] => {
             for (key, value) in kwargs.by_ref() {
-                if key.py_eq(str_ignore_dst)? {
-                    got_ignore_dst = true;
-                } else if key.py_eq(str_stale_offset_ok)? {
+                if key.py_eq(str_stale_offset_ok)? {
                     suppress_stale = value.is_truthy();
                 } else {
                     raise_type_err(format!(
@@ -1029,13 +972,6 @@ fn shift_method(
             }
             if let Some(t) = arg.extract(time_delta_type) {
                 tdelta = t;
-            } else if let Some(ddelta) = arg.extract(date_delta_type) {
-                months = ddelta.months;
-                days = ddelta.days;
-            } else if let Some(dt) = arg.extract(datetime_delta_type) {
-                months = dt.ddelta.months;
-                days = dt.ddelta.days;
-                tdelta = dt.tdelta;
             } else if let Some(d) = arg.extract(itemized_date_delta_type) {
                 let (m, dy) = d.to_months_days().ok_or_range_err()?;
                 months = m;
@@ -1052,10 +988,7 @@ fn shift_method(
         [] => {
             let mut units = DeltaUnitSet::EMPTY;
             handle_kwargs(fname, kwargs, |key, value, eq| {
-                if eq(key, str_ignore_dst) {
-                    got_ignore_dst = true;
-                    Ok(true)
-                } else if eq(key, str_stale_offset_ok) {
+                if eq(key, str_stale_offset_ok) {
                     suppress_stale = value.is_truthy();
                     Ok(true)
                 } else {
@@ -1088,9 +1021,6 @@ fn shift_method(
         ))?,
     }
 
-    if got_ignore_dst {
-        warn_with_class(warn_deprecation, doc::IGNORE_DST_DEPRECATED_MSG, 1)?;
-    }
     if !suppress_stale {
         offset_stale_warning(state, doc::OFFSET_SHIFT_STALE_MSG)?;
     }
@@ -1153,7 +1083,7 @@ fn __reduce__(cls: HeapType<OffsetDateTime>, slf: OffsetDateTime) -> PyResult<Ow
         .into_pytuple()
 }
 
-/// checks the args comply with (ts: ?, /, *, offset: ?, ignore_dst: ?, stale_offset_ok: ?)
+/// checks the args comply with (ts: ?, /, *, offset: ?, stale_offset_ok: ?)
 fn check_from_timestamp_args_return_offset(
     fname: &str,
     args: &[PyObj],
@@ -1161,13 +1091,11 @@ fn check_from_timestamp_args_return_offset(
     state: &State,
 ) -> PyResult<Offset> {
     let &State {
-        str_ignore_dst,
         str_stale_offset_ok,
         str_offset,
         time_delta_type,
         ..
     } = state;
-    let mut got_ignore_dst = false;
     let mut suppress_stale = false;
     let mut offset = None;
     if args.len() != 1 {
@@ -1179,9 +1107,7 @@ fn check_from_timestamp_args_return_offset(
     }
 
     handle_kwargs("from_timestamp", kwargs, |key, value, eq| {
-        if eq(key, str_ignore_dst) {
-            got_ignore_dst = true;
-        } else if eq(key, str_stale_offset_ok) {
+        if eq(key, str_stale_offset_ok) {
             suppress_stale = value.is_truthy();
         } else if eq(key, str_offset) {
             offset = Some(Offset::from_obj(value, time_delta_type)?);
@@ -1191,9 +1117,6 @@ fn check_from_timestamp_args_return_offset(
         Ok(true)
     })?;
 
-    if got_ignore_dst {
-        warn_with_class(state.warn_deprecation, doc::IGNORE_DST_DEPRECATED_MSG, 1)?;
-    }
     if !suppress_stale {
         offset_stale_warning(state, doc::OFFSET_FROM_TIMESTAMP_STALE_MSG)?;
     }
@@ -1262,38 +1185,6 @@ fn from_timestamp_nanos(
     .to_obj(cls)
 }
 
-fn parse_strptime(
-    cls: HeapType<OffsetDateTime>,
-    args: &[PyObj],
-    kwargs: &mut IterKwargs,
-) -> PyReturn {
-    let state = cls.state();
-    warn_with_class(
-        state.warn_deprecation,
-        c"parse_strptime() is deprecated; use parse() with a format pattern instead.",
-        1,
-    )?;
-    let format_obj = match kwargs.next() {
-        Some((key, value)) if kwargs.len() == 1 && key.py_eq(state.str_format)? => value,
-        _ => raise_type_err("parse_strptime() requires exactly one keyword argument `format`")?,
-    };
-    let &[arg_obj] = args else {
-        raise_type_err(format!(
-            "parse_strptime() takes exactly 1 positional argument, got {}",
-            args.len()
-        ))?
-    };
-
-    let args = (arg_obj.newref(), format_obj.newref()).into_pytuple()?;
-    let parsed = state
-        .strptime
-        .call(*args)?
-        .cast_exact::<PyDateTime>()
-        .ok_or_type_err("strptime() returned non-datetime")?;
-
-    OffsetDateTime::from_py(*parsed)?.to_obj(cls)
-}
-
 fn format_rfc2822(_: PyType, slf: OffsetDateTime) -> PyReturn {
     let fmt = rfc2822::format(slf);
     // SAFETY: we know the format is ASCII only
@@ -1321,12 +1212,8 @@ fn round(
     let round::Args {
         increment,
         mode,
-        got_ignore_dst,
         suppress_stale,
     } = round::Args::parse(state, args, kwargs, true)?;
-    if got_ignore_dst {
-        warn_with_class(state.warn_deprecation, doc::IGNORE_DST_DEPRECATED_MSG, 1)?;
-    }
     if !suppress_stale {
         offset_stale_warning(state, doc::OFFSET_ROUND_STALE_MSG)?;
     }
@@ -1579,16 +1466,6 @@ static mut METHODS: &[PyMethodDef] = &[
     method0!(OffsetDateTime, to_stdlib, doc::BASICCONVERSIONS_TO_STDLIB),
     method0!(
         OffsetDateTime,
-        py_datetime,
-        doc::BASICCONVERSIONS_PY_DATETIME
-    ),
-    classmethod1!(
-        OffsetDateTime,
-        from_py_datetime,
-        doc::BASICCONVERSIONS_FROM_PY_DATETIME
-    ),
-    method0!(
-        OffsetDateTime,
         to_instant,
         doc::EXACTANDLOCALTIME_TO_INSTANT
     ),
@@ -1657,11 +1534,6 @@ static mut METHODS: &[PyMethodDef] = &[
         OffsetDateTime,
         replace_time,
         doc::OFFSETDATETIME_REPLACE_TIME
-    ),
-    classmethod_kwargs!(
-        OffsetDateTime,
-        parse_strptime,
-        doc::OFFSETDATETIME_PARSE_STRPTIME
     ),
     method_kwargs!(OffsetDateTime, add, doc::OFFSETDATETIME_ADD),
     method_kwargs!(OffsetDateTime, subtract, doc::OFFSETDATETIME_SUBTRACT),