+++ tests/test_itemized_date_delta.py
@@ -6,15 +6,11 @@ from typing import Any, Literal, Sequence, cast
 import pytest
 from whenever import Date, ItemizedDateDelta, ItemizedDelta
 
-from .common import AlwaysEqual, NeverEqual
-from .test_date_delta import INVALID_DDELTAS
+from .common import INVALID_DDELTAS, AlwaysEqual, NeverEqual
 
 UNITS = cast(
     Sequence[Literal["years", "months", "weeks", "days"]],
     "years months weeks days".split(),
-)
-pytestmark = pytest.mark.filterwarnings(
-    "ignore::whenever.WheneverDeprecationWarning"
 )